fix: Address PR #37 review feedback on JSON Schema list-type handling
- Extended JSON Schema array notation fix to langchain_adapter.py
- Added support for both array notation ("type": ["string", "number"])
  and anyOf structures in LangChain adapter
- Reorganized tests per reviewer feedback:
  * Moved simplified direct test to tests/utils/test_modeling.py
  * Added E2E test to test_langchain_adapter.py
  * Removed redundant test file

This commit ensures consistent handling of JSON Schema list-type fields
across all adapters and addresses all feedback from the PR review.
