+++ tests/test_langchain_adapter.py
@@ -8,6 +8,32 @@ from mcpadapt.langchain_adapter import LangChainAdapter
 
 
 @pytest.fixture
+def json_schema_array_type_server_script():
+    """
+    Create a server with a tool that uses array notation for type fields.
+    This tests handling of JSON Schema 'type': ['string', 'number'] syntax.
+    """
+    return dedent(
+        '''
+        from mcp.server.fastmcp import FastMCP
+
+        mcp = FastMCP("JSON Schema Array Type Test Server")
+
+        @mcp.tool()
+        def multi_type_tool(
+            id: str | int,  # This becomes {"type": ["string", "number"]} in JSON Schema
+            name: str | None = None,  # Tests nullable with array type
+        ) -> str:
+            """Test tool with a parameter that accepts multiple types using array notation"""
+            id_type = type(id).__name__
+            return f"Received ID: {id} (type: {id_type}), Name: {name}"
+        
+        mcp.run()
+        '''
+    )
+
+
+@pytest.fixture
 def echo_server_script():
     return dedent(
         '''
@@ -110,6 +136,31 @@ def test_basic_sync_sse(echo_sse_server):
         LangChainAdapter(),
     ) as tools:
         assert len(tools) == 1
+
+
+def test_json_schema_array_type_handling(json_schema_array_type_server_script):
+    """
+    Test that MCPAdapt correctly handles JSON Schema with array notation for types.
+    This ensures our fix for 'unhashable type: list' error is working.
+    """
+    with MCPAdapt(
+        StdioServerParameters(
+            command="uv",
+            args=["run", "python", "-c", json_schema_array_type_server_script],
+        ),
+        LangChainAdapter(),
+    ) as tools:
+        # Verify the tool was successfully loaded
+        assert len(tools) == 1
+        assert tools[0].name == "multi_type_tool"
+
+        # Test with string type
+        result_string = tools[0].invoke({"id": "abc123", "name": "test"})
+        assert "Received ID: abc123 (type: str)" in result_string
+
+        # Test with integer type
+        result_int = tools[0].invoke({"id": 42, "name": "test"})
+        assert "Received ID: 42 (type: int)" in result_int
 
 
 def test_tool_name_with_dashes():
